+++ packages/icons/src/__generated__/ui/data/names.ts
@@ -281,6 +281,8 @@ const names: UiIconName[] = [
   'login',
   'sortDownCenterInactive',
   'sortUpCenterInactive',
+  'lineChartCrypto',
+  'cashCoins',
 ];
 
 export default names;