+++ packages/illustrations/CHANGELOG.md
@@ -8,6 +8,21 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 3.5.0 (4/25/2024 PST)
+
+#### 🚀 Updates
+
+- Publish illustrations 4/25/2024. [[#2968](https://github.cbhq.net/frontend/cds/pull/2968)]
+
+### 🚀 Added
+
+Generated with `yarn nx run illustrations:sync`
+
+#### Pictogram
+
+- developerSDKNavigation
+- verifiedPools
+
 ## 3.4.0 (4/18/2024 PST)
 
 #### 🚀 Updates